+++ test/fixtures/invalid_task_config.exs
@@ -1,0 +1,29 @@
+[
+  version: "0.7.0",
+  # Can also be set/reset with `--autocorrect`/`--no-autocorrect`.
+  autocorrect: true,
+  # With "--dry" no changes will be written to the files.
+  # Can also be set/reset with `--dry`/`--no-dry`.
+  # If dry is true then verbose is also active.
+  dry: false,
+  # Can also be set/reset with `--verbose`/`--no-verbose`.
+  verbose: false,
+  # Can be overwritten by calling `mix recode "lib/**/*.ex"`.
+  inputs: ["{mix,.formatter}.exs", "{apps,config,lib,test}/**/*.{ex,exs}"],
+  formatters: [Recode.CLIFormatter],
+  tasks: [
+    # Tasks could be added by a tuple of the tasks module name and an options
+    # keyword list. A task can be deactivated by `active: false`. The execution of
+    # a deactivated task can be forced by calling `mix recode --task ModuleName`.
+    {Recode.Task.AliasExpansion, []},
+    {Recode.Task.AliasOrder, []},
+    {Recode.Task.Dbg, [autocorrect: false, invalid: :key]},
+    {Recode.Task.EnforceLineLength, [active: false]},
+    {Recode.Task.FilterCount, []},
+    {Recode.Task.PipeFunOne, []},
+    {Recode.Task.SinglePipe, []},
+    {Recode.Task.Specs, [exclude: "test/**/*.{ex,exs}", config: [only: :visible]]},
+    {Recode.Task.TestFileExt, []},
+    {Recode.Task.UnusedVariable, [active: false]}
+  ]
+]